refactor: move getRandomInt file
